Refactor: Move the WEB3 context from hook to context folder
