+++ app/src/components/comments/index.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react'
 import styled from 'styled-components'
 
 import { THREEBOX_ADMIN_ADDRESS, THREEBOX_SPACE_NAME } from '../../common/constants'
-import { ConnectedWeb3Context, useConnectedWeb3Context } from '../../hooks/connectedWeb3'
+import { ConnectedWeb3Context, useConnectedWeb3Context } from '../../contexts/connectedWeb3'
 
 const CommentsTitle = styled.h3`
   color: #000;
