+++ app/src/components/market/common/customizable_tokens_select/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
 
-import { ConnectedWeb3Context } from '../../../../hooks/connectedWeb3'
+import { ConnectedWeb3Context } from '../../../../contexts/connectedWeb3'
 import { Token } from '../../../../util/types'
 import { FormRowLink } from '../../../common/form/form_row_link'
 import { Tokens } from '../tokens'
