+++ app/src/components/market/common/customizable_tokens_select/modal_collateral.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent, HTMLAttributes, useState } from 'react'
 import styled from 'styled-components'
 
+import { ConnectedWeb3Context } from '../../../../contexts/connectedWeb3'
 import { useCollateral } from '../../../../hooks'
-import { ConnectedWeb3Context } from '../../../../hooks/connectedWeb3'
 import { Token } from '../../../../util/types'
 import { Button } from '../../../button'
 import { ButtonType } from '../../../button/button_styling_types'
