+++ app/src/components/market/common/tokens/index.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 
 import { DEFAULT_TOKEN } from '../../../../common/constants'
+import { ConnectedWeb3Context } from '../../../../contexts/connectedWeb3'
 import { useContracts } from '../../../../hooks'
-import { ConnectedWeb3Context } from '../../../../hooks/connectedWeb3'
 import { getToken } from '../../../../util/networks'
 import { Token } from '../../../../util/types'
 import { Select } from '../../../common/form/select'
