+++ app/src/components/market/routes/market_routes.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import { Redirect, RouteComponentProps } from 'react-router'
 
 import { FETCH_DETAILS_INTERVAL } from '../../../common/constants'
+import { useConnectedWeb3Context } from '../../../contexts/connectedWeb3'
 import { useCheckContractExists, useMarketMakerData } from '../../../hooks'
-import { useConnectedWeb3Context } from '../../../hooks/connectedWeb3'
 import { MarketDetailsPage } from '../../../pages'
 import { getLogger } from '../../../util/logger'
 import { isAddress } from '../../../util/tools'
