+++ app/src/components/market/sections/market_create/steps/items/ask_question_step.tsx
@@ -5,7 +5,7 @@ import { useHistory } from 'react-router'
 import styled, { css } from 'styled-components'
 
 import { DOCUMENT_VALIDITY_RULES, MAX_OUTCOME_ALLOWED } from '../../../../../../common/constants'
-import { useConnectedWeb3Context } from '../../../../../../hooks/connectedWeb3'
+import { useConnectedWeb3Context } from '../../../../../../contexts/connectedWeb3'
 import { Arbitrator, FormState, Question } from '../../../../../../util/types'
 import { Button, ButtonContainer } from '../../../../../button'
 import { ButtonType } from '../../../../../button/button_styling_types'
