+++ app/src/components/market/sections/market_create/steps/items/import_market_content.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent, HTMLAttributes, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
 
+import { ConnectedWeb3Context } from '../../../../../../contexts/connectedWeb3'
 import { useAsyncDerivedValue, useContracts, useMarketMakerData } from '../../../../../../hooks'
-import { ConnectedWeb3Context } from '../../../../../../hooks/connectedWeb3'
 import { useGraphMarketsFromQuestion } from '../../../../../../hooks/useGraphMarketsFromQuestion'
 import { useRealityLink } from '../../../../../../hooks/useRealityLink'
 import { getArbitratorFromAddress } from '../../../../../../util/networks'
