+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -6,8 +6,8 @@ import { RouteComponentProps, useHistory, useLocation, withRouter } from 'react-
 import styled from 'styled-components'
 
 import { STANDARD_DECIMALS } from '../../../../../common/constants'
+import { WhenConnected, useConnectedWeb3Context } from '../../../../../contexts/connectedWeb3'
 import { useContracts, useGraphMarketUserTxData, useSymbol } from '../../../../../hooks'
-import { WhenConnected, useConnectedWeb3Context } from '../../../../../hooks/connectedWeb3'
 import { ERC20Service, RealitioService } from '../../../../../services'
 import { CompoundService } from '../../../../../services/compound_service'
 import { getLogger } from '../../../../../util/logger'
