+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -3,8 +3,8 @@ import React, { useEffect, useState } from 'react'
 import { RouteComponentProps, useHistory, useLocation, withRouter } from 'react-router-dom'
 import styled from 'styled-components'
 
+import { WhenConnected, useConnectedWeb3Context } from '../../../../../contexts/connectedWeb3'
 import { useCompoundService, useGraphMarketUserTxData } from '../../../../../hooks'
-import { WhenConnected, useConnectedWeb3Context } from '../../../../../hooks/connectedWeb3'
 import { useRealityLink } from '../../../../../hooks/useRealityLink'
 import { getNativeAsset, getToken, networkIds } from '../../../../../util/networks'
 import { getSharesInBaseToken, getUnit, isDust } from '../../../../../util/tools'
